+++ .storybook/preview.js
@@ -1,8 +1,15 @@
 // import '!style-loader!css-loader!sass-loader!../src/assets/scss/main.scss';
 import '../src/assets/css/main.css';
-  
+
+import Vuex from 'vuex';
+import Vue from 'vue';
+import store from '../src/components/store';
+
+Vue.use(Vuex);
+Vue.prototype.$store = store;
+
 export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
+  actions: { argTypesRegex: '^on[A-Z].*' },
   controls: {
     matchers: {
       color: /(background|color)$/i,
@@ -11,13 +18,7 @@ export const parameters = {
   },
   options: {
     storySort: {
-      order: [
-        'About PHOTON',
-        'Global',
-        'Button & Tags',
-        'Forms',
-        'Components',
-      ], 
+      order: ['About PHOTON', 'Global', 'Button & Tags', 'Forms', 'Components'],
     },
   },
-}
+};