+++ src/components/ListItem/ListItem.vue
@@ -6,7 +6,7 @@
   </li>
 </template>
 <script lang="ts">
-import Vue, { PropType } from 'vue';
+import Vue from 'vue';
 
 export default Vue.extend({
   name: 'PLi',