+++ CHANGELOG.md
@@ -5,17 +5,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.635] - 2022-01-14
+### Add
+- Husky git precommit
+
 ## [0.1.634] - 2022-01-13
 ### Fix
 - InputPhone is have an extra empty label cause of conditional 
 
-## [0.1.632] - 2022-01-11
+## [0.1.634] - 2022-01-12
+### Add
+- SideBar Component
+
+## [0.1.633] - 2022-01-11
 ### Fix
 - InfoCard remove red background of splitter
 
-## [0.1.633] - 2022-01-10
-### Add
-- Husky git precommit
+## [0.1.634] - 2021-12-21
+### Fix
+- Added Vuex to Toast Component
+- Toast & Message unit tests
 
 ## [0.1.632] - 2022-01-07
 ### Fix